Updated mock unit tests and closed all opened sockets in test cases make test and make test-dry both behave correctly now
