+++ tests/test_Input.py
@@ -8,7 +8,7 @@ if int(os.environ['MC_Live_Connection'])==1:
     from mcpi.minecraft import Minecraft
     print("importing real MC")
 else:
-    from .fake_minecraft import Minecraft
+    from .mock_minecraft import Minecraft
     print("importing fame MC")
 
 import pickle
@@ -34,12 +34,13 @@ class TestInput(unittest.TestCase):
         server_ip, server_port = pickle.load(my_file)
         my_file.close()
         mc = Minecraft.create(server_ip,server_port)
+        
         my_controller = input.TactileInput([mc],start_coord=start_coord, end_coord=end_coord)
         self.assertIsInstance(my_controller, input.TactileInput)
 
         my_controller = input.RangeInput([mc],start_coord=start_coord, end_coord=end_coord)
         self.assertIsInstance(my_controller, input.RangeInput)
-
+        mc.conn.socket.close()
     def testInputBlockLineX(self):
 
         """
@@ -59,7 +60,7 @@ class TestInput(unittest.TestCase):
 
         my_controller = input.RangeInput([mc],start_coord=start_coord, end_coord=end_coord)
         self.assertIsInstance(my_controller, input.RangeInput)
-
+        mc.conn.socket.close()
     def testInputBlockLineZ(self):
 
         """
@@ -79,7 +80,7 @@ class TestInput(unittest.TestCase):
 
         my_controller = input.RangeInput([mc],start_coord=start_coord, end_coord=end_coord)
         self.assertIsInstance(my_controller, input.RangeInput)
-
+        mc.conn.socket.close()
     def testInputPlaneXZ(self):
 
         """
@@ -99,7 +100,7 @@ class TestInput(unittest.TestCase):
         
         with self.assertRaises(RuntimeError):
             input.RangeInput([mc],start_coord=start_coord, end_coord=end_coord)
-
+        mc.conn.socket.close()
 
     def testInputBlockLineY(self):
 
@@ -120,6 +121,7 @@ class TestInput(unittest.TestCase):
         
         with self.assertRaises(RuntimeError):
             input.RangeInput([mc],start_coord=start_coord, end_coord=end_coord) 
+        mc.conn.socket.close()
 
     def testInputCuboid(self):
 
@@ -140,38 +142,7 @@ class TestInput(unittest.TestCase):
         
         with self.assertRaises(RuntimeError):
             input.RangeInput([mc],start_coord=start_coord, end_coord=end_coord)  
-
-    def testRangeInputInstance(self):
-        """
-        Tests for RangeInput defining a line of blocks and testing for a player block at the beginning, middle, and end of that line to check linear interpolation values (0-1)
-        """
-
-        #player_id = 0  #just need some integer here
-        start_coord = MCVector.from_MCWorld_Vec(vec3.Vec3(39536,83,39955))
-        end_coord = MCVector.from_MCWorld_Vec(vec3.Vec3(39536,83,39962))
-
-        my_file= open( "server.pkl", "rb" ) 
-        server_ip, server_port = pickle.load(my_file)
-        my_file.close()
-        mc = Minecraft.create(server_ip,server_port)
-        
-        mc.entity.player_pos=MCVector.from_MCWorld_Vec(vec3.Vec3(39536,84,39955)) # player at start of platform
-        my_controller = input.RangeInput([mc],start_coord=start_coord, end_coord=end_coord)
-        my_controller.scanInput()
-        lerp_val=my_controller.getInputValue()
-        self.assertAlmostEqual(lerp_val, 0.0)
-
-        mc.entity.player_pos=MCVector.from_MCWorld_Vec(vec3.Vec3(39536,84,39962)) # player at end of platform
-        my_controller.scanInput()
-        lerp_val=my_controller.getInputValue()
-        self.assertAlmostEqual(lerp_val, 1.0)
-
-        mc.entity.player_pos=MCVector.from_MCWorld_Vec(vec3.Vec3(39536,84,39958)) # player somewhere near middle of controller
-        my_controller.scanInput()
-        lerp_val=my_controller.getInputValue()
-        self.assertAlmostEqual(lerp_val, 0.42857142)
-                
-
+        mc.conn.socket.close()
 
 
 if __name__ == '__main__':
